feat: add working block jacobi

+++ src/cpp/main.cpp
@@ -31,8 +31,8 @@ int main(int argc, char **argv) {
     run_all_ode_experiments(Half);
   } else if (experiment == "testing") {
     // run_poisson_equation_experiments(Double);
-    run_poisson_equation_experiments(Single);
-    // run_poisson_equation_experiments(Half);
+    // run_poisson_equation_experiments(Single);
+    run_poisson_equation_experiments(Half);
   }
   return 0;
 }
